# Add commitment cache optimization