--- a/src/setup/settings.h
+++ b/src/setup/settings.h
@@ -19,6 +19,8 @@
 #include "common/ec.h"
 #include "common/fr.h"
 
+#include <stdatomic.h> /* For atomic_flag */
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // Types
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -76,4 +78,12 @@ typedef struct {
     size_t wbits;
     /** The scratch size for the fixed-base MSM. */
     size_t scratch_size;
+    /** A lock for the commitment cache. */
+    atomic_flag *comm_cache_lock;
+    /** The bytes used as the commitment cache key. */
+    uint8_t *comm_cache_key;
+    /** The bytes used as the commitment cache value. */
+    uint8_t *comm_cache_value;
+    /** The max number of commitments the cache supports. */
+    uint64_t comm_cache_size;
 } KZGSettings;