--- a/src/setup/settings.h
+++ b/src/setup/settings.h
@@ -77,7 +77,7 @@ typedef struct {
     /** The scratch size for the fixed-base MSM. */
     size_t scratch_size;
     /** A lock for the commitment cache. */
-    uint8_t *comm_cache_lock;
+    long *comm_cache_lock;
     /** The bytes used as the commitment cache key. */
     uint8_t *comm_cache_key;
     /** The bytes used as the commitment cache value. */